+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 * Save mod list changes immediately rather than on pressing Play
 * Display mod dependencies by name when not in Modder Mode
 * Add default value options to SpireConfig get methods
+* Add setting to enable achievements
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)